Refactor of GET and POST routes on category
Restructuring the controller to follow REST conventions as well as
putting some logging into the methods for aid in debugging in the cloud
envirement.

Watch for the active property on the category field must be pased in
using active not isActive. Because of this I have added an explicit
JsonValue mapping.